requirements.txt
feedparser
newspaper3k
schedule
nltk
